refactor: moved App.css; chore: made issue file for pending changes
