+++ frontend/.gitignore
@@ -2,6 +2,6 @@ node_modules
 src/api_key.js
 nohup.out
 package-lock.json
-
+issues.txt
 src/assets/concepts/
 formatTempPlan.js
