+++ frontend/src/App.css
@@ -1,0 +1,228 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+.App {
+	text-align: center;
+}
+
+.App-logo {
+	height: 40vmin;
+	pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+	.App-logo {
+		animation: App-logo-spin infinite 20s linear;
+	}
+}
+
+.App-header {
+	background-color: #282c34;
+	min-height: 100vh;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	font-size: calc(10px + 2vmin);
+	color: white;
+}
+
+.App-link {
+	color: #61dafb;
+}
+
+@keyframes App-logo-spin {
+	from {
+		transform: rotate(0deg);
+	}
+	to {
+		transform: rotate(360deg);
+	}
+}
+
+
+
+/* === LOGIN ==== */
+body {
+	margin: 0;
+	padding: 0;
+	box-sizing: border-box;
+	/* background: linear-gradient(to right,  #00153a, #f0f0f5); */
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	min-height: 100vh;
+	font-family: Arial, sans-serif;
+}
+
+/* Authentication Form Container */
+.center {
+	background: rgba(255, 255, 255, 0.8);
+	margin: auto 1rem;
+	padding: 2.5rem 4rem;
+	border-radius: 0.625rem;
+	width: 90%;
+	max-width: 25rem;
+	box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
+    /* text-align: center; */
+}
+
+
+.center h1 {
+	font-size: 2.25rem;
+    margin: 0.625rem 0;
+    color: #002366;
+	text-align: center;
+}
+
+form {
+	padding: 1rem;
+}
+
+.user__input {
+	width: 100%;
+}
+
+.center label {
+    font-size: 1em;
+    margin: 0.5rem 0;
+    color: #333;
+}
+
+/* Input Fields */
+.center input[type="text"],
+.center input[type="email"],
+.center input[type="password"] {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+    width: 90%;
+    padding: 0.75rem;
+    margin: 0.5rem 0;
+    border: 0.0625rem solid #ccc;
+    border-radius: 0.3125rem;
+    font-size: 1rem;
+}
+
+/* Authentication Submit Button */
+.center button {
+    background-color: #28a745;
+    color: #fff;
+    padding: 12px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 1em;
+    width: 100%;
+    transition: box-shadow 0.3s;
+}
+
+/* Glowing effect on button */
+.center button:hover {
+    box-shadow: 0 0 15px rgba(40, 167, 69, 0.6);
+}
+
+/* Text styling for links */
+.center p {
+    font-size: 0.9rem;
+    margin: 1.25rem 0;
+}
+
+.center a {
+    color: #5c5c5c;
+    text-decoration: none;
+}
+
+.center a:hover {
+    text-decoration: underline;
+}
+
+.register, .login {
+	margin: 1.5rem 0;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+.register a, .login a {
+	margin-left: 0.25rem;
+	cursor: pointer;
+}
+.container {
+	background-color: white;
+	padding: 20px;
+	border-radius: 8px;
+	box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
+	width: 80%;
+	max-width: 400px;
+	/* text-align: left; */
+}
+
+.advice-section {
+	text-align: center;
+	margin-top: 20px;
+	margin-bottom: 30px;
+}
+
+h3, h4 {
+	margin: 10px 0;
+}
+
+#answer-container {
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	margin: 20px 0;
+}
+
+input[type="radio"], input[type="checkbox"] {
+	margin-right: 10px;
+}
+
+label {
+	display: flex;
+	align-items: center;
+	margin-bottom: 10px;
+}
+
+button {
+	background-color: #4CAF50;
+	color: white;
+	padding: 10px 20px;
+	border: none;
+	border-radius: 5px;
+	cursor: pointer;
+	margin-top: 10px;
+}
+
+button:hover {
+	background-color: #45a049;
+}
+
+.revise-plan {
+	margin-top: 20px;
+}
+
+.revise-plan input {
+	width: calc(100% - 22px);
+	padding: 10px;
+	margin-top: 10px;
+	border: 1px solid #ccc;
+	border-radius: 4px;
+}
+
+.revised-plan {
+	margin-top: 20px;
+	border-top: 1px solid #ccc;
+	padding-top: 10px;
+}
+
+.revised-plan h4 {
+	margin: 10px 0;
+}
+
+/* Error message */
+.error {
+	color: red;
+}
