+++ frontend/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css'
-// import App from './App';
+import './index.css';
+import './App.css';
 import Routes from './routes';
 import reportWebVitals from './reportWebVitals';
