+++ frontend/src/pages/IntroPage.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import '../App.css';
 import { useNavigate } from 'react-router-dom';
 
 const IntroPage = () => {
