+++ frontend/src/pages/QuizPage.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import '../App.css';
 import useSurveyData from '../hooks/useSurveyData.js';
 import QuestionCard from '../components/QuestionCard.jsx';
 import PlanDisplay from '../components/PlanDisplay.jsx';
