+++ frontend/src/pages/Register.jsx
@@ -1,4 +1,4 @@
-import '../styles/App.css';
+import '../App.css';
 // import axios from "axios";
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
